extract monaco into it's own component, remove definitions context
